Issue 75 - Importing trac ticket with missing attachment breaks the import process
If for some reason an attachment in the trac database is missing, this throws an
error during the import.  This should not halt the import process.  This fix checks
for the missing attachment error message and does not exit, and it also correctly
handles the missing attachment in other parts of the import code.

https://pagure.io/pagure-importer/issue/75