Make to_timestamp a function
to_timestamp has nothing to do with the TracImporter class it was on, the proof is that
it never used the ``self`` argument passed. So this commit moves it to a function instead
of it being a method of TracImporter.

Eventually, it could even be moved into another module as it is also not specific to trac
and therefore, could eventually be used by other pieces of code.